services/ai-threat-analyzer/requirements.txt
fastapi
uvicorn
pika
pymongo
joblib
python-dotenv
